pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepfold"
version = "0.1.0"
description = "A Django project for managing publishing outlets"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "Django==4.1",
    "mysqlclient",
    "dj-database-url",
    "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "app"}

[tool.setuptools.packages.find]
where = ["app"]